pyproject.toml
[project]
name = "bulletin-scraper"
version = "0.1.0"
description = "Modular bulletin ingestion and LangGraph-based update workflows for church schedule, events, and intentions data."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "cloudscraper>=1.2.71",
    "langchain-core>=0.3.79",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.7",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "pymupdf>=1.26.5",
    "pillow>=11.3.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[project.scripts]
bulletin-scraper = "bulletin_scraper.cli:main"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]